% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/animals.R
\name{animals}
\alias{animals}
\title{#Create a dataset for classification}
\usage{
animals(rho, sdXY, muXY, noSim, seed = 5341)
}
\arguments{
\item{rho}{A vector of 3 numbers belonging to \code{[-1,1]} which indicate the
correlation between two features of the three classes to simulate.}

\item{sdXY}{A list of 3 vectors which indicate the standard deviation of features
of a class.}

\item{muXY}{A list of 3 vectors which indicate the mean of features
of a class.}

\item{noSim}{A vector of 3 numbers of no of simulations desired for each class}

\item{seed}{Specifies seed. Default Value = 1234}
}
\value{
Data Frame of simulations with their classes and coded dummy variables
}
\description{
#Create a dataset for classification
}
\examples{
noAnimals <- c(200, 200, 200)
rho  <- c(-0.2, 0.8, 0.02)
sdXY <- list(c(2, 8), c(3.5, 12), c(1, 3))
muXY <- list(c(6, 30), c(12, 50), c(4,10))
animalsDF <- animals(rho, sdXY, muXY, noAnimals)
}

