% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genBVN.R
\name{genBVN}
\alias{genBVN}
\title{Creates realizations of a normal Bi-variate random variable}
\usage{
genBVN(n = 1, seed = 1234, muXY = c(0, 1), sigmaXY = diag(2))
}
\arguments{
\item{n}{Number of realizations.}

\item{seed}{Set seed.}

\item{muXY}{Vector of two numbers that represent means of features.}

\item{sigmaXY}{Variance Covariance Matrix}
}
\value{
Realizations of n bivariate normal random variable with means muXY and
variance covariance sigmaXY.
}
\description{
Creates realizations of a normal Bi-variate random variable
}
\examples{
rho  <- -0.2
sdX <- 2
sdY <- 8
muXY <- c(0,1)
sigmaXY <- sigmaXY(rho, sdX, sdY)
genBVN (n = 1, seed = 1234, muXY, sigmaXY)
}

