% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getPrediction.R
\name{getPrediction}
\alias{getPrediction}
\title{Classifies data points to ine of three animal classes}
\usage{
getPrediction(animalsDF)
}
\arguments{
\item{animalsDF}{A data frame of simulated points representing three classes
along with their classes and dummy categorical variables corresponding to each class.}
}
\value{
List of predicted labels along with regression information of the three animal classes
}
\description{
Function takes input of points and classifies them into one of three classes:
Cats, Dogs and Maggots.
}
\examples{
noAnimals <- c(200, 200, 200)
rho  <- c(-0.2, 0.8, 0.02)
sdXY <- list(c(2, 8), c(3.5, 12), c(1, 3))
muXY <- list(c(6, 30), c(12, 50), c(4,10))
animalsDF <- animals(rho, sdXY, muXY, noAnimals)
predictionList <- getPrediction(animalsDF)
}

