% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/numToChar.R
\name{numToChar}
\alias{numToChar}
\title{Classify k-nearest neighbour vector}
\usage{
numToChar(classVector)
}
\arguments{
\item{classVector}{A vector of numerical class predictions for new data points.}
}
\value{
Vector of character labels of classes
}
\description{
A numerical vector of probabilities of a point belonging to a class of cats
dogs or maggots. The output is the character label of the class with the highest
probability
}
\examples{
num.class <- c(.73, 0.5, 0.98)
numToChar(num.class)
}

