% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sigmaXY.R
\name{sigmaXY}
\alias{sigmaXY}
\title{Creates a Variance Covariance Matrix}
\usage{
sigmaXY(rho, sdX, sdY)
}
\arguments{
\item{rho}{A number belonging to \code{[-1,1]} which indicate the
correlation between two features of the three classes to simulate.}

\item{sdX}{Standard deviaton of first feature}

\item{sdY}{Standard deviaton of first feature}
}
\value{
Variance Covariance matrix
}
\description{
Given correlations and standard deviations of both features the function returns
a variance covariance matrix
}
\examples{
rho  <- -0.2
sdX <- 2
sdY <- 10
sigmaXY(rho, sdX, sdY)
}

